solution/requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
boto
boto3
